Created separated IndentCodeBlockCommand and OutdentCodeBlockCommand classes.
