+++ src/utils.js
@@ -7,6 +7,8 @@
  * @module code-block/utils
  */
 
+import first from '@ckeditor/ckeditor5-utils/src/first';
+
 /**
  * Returns code block languages as defined in `config.codeBlock.languages` but processed to consider
  * the editor localization, i.e. to display {@link module:code-block/codeblock~CodeBlockLanguageDefinition}
@@ -77,4 +79,85 @@ export function rawSnippetTextToModelDocumentFragment( writer, text ) {
 	}
 
 	return fragment;
+}
+
+/**
+ * Returns an array of all model positions within the selection that represent code block lines.
+ *
+ * If the selection is collapsed, it returns the exact selection anchor position:
+ *
+ *		<codeBlock>[]foo</codeBlock>        ->     <codeBlock>^foo</codeBlock>
+ *		<codeBlock>foo[]bar</codeBlock>     ->     <codeBlock>foo^bar</codeBlock>
+ *
+ * Otherwise, it returns positions **before** each text node belonging to all code blocks contained by the selection:
+ *
+ *		<codeBlock>                                <codeBlock>
+ *		    foo[bar                                   ^foobar
+ *		    <softBreak></softBreak>         ->        <softBreak></softBreak>
+ *		    baz]qux                                   ^bazqux
+ *		</codeBlock>                               </codeBlock>
+ *
+ * it also works across other non–code blocks:
+ *
+ *		<codeBlock>                                <codeBlock>
+ *		    foo[bar                                   ^foobar
+ *		</codeBlock>                               </codeBlock>
+ *		<paragraph>text</paragraph>         ->     <paragraph>text</paragraph>
+ *		<codeBlock>                                <codeBlock>
+ *		    baz]qux                                   ^bazqux
+ *		</codeBlock>                               </codeBlock>
+ *
+ * **Note:** The positions are in the reverse order so they do not get outdated when iterating over them and
+ * the writer inserts or removes things.
+ *
+ * **Note:** The position is situated after the leading white spaces in the text node.
+ *
+ * @param {module:engine/model/model~Model} model
+ * @returns {Array.<module:engine/model/position~Position>}
+ */
+export function getIndentOutdentPositions( model ) {
+	const selection = model.document.selection;
+	const positions = [];
+
+	// When the selection is collapsed, there's only one position we can indent or outdent.
+	if ( selection.isCollapsed ) {
+		positions.push( selection.anchor );
+	}
+
+	// When the selection is NOT collapsed, collect all positions starting before text nodes
+	// (code lines) in any <codeBlock> within the selection.
+	else {
+		// Walk backward so positions we're about to collect here do not get outdated when
+		// inserting or deleting using the writer.
+		const walker = selection.getFirstRange().getWalker( {
+			ignoreElementEnd: true,
+			direction: 'backward'
+		} );
+
+		for ( const { item } of walker ) {
+			if ( item.is( 'textProxy' ) && item.parent.is( 'codeBlock' ) ) {
+				const leadingWhiteSpaces = getLeadingWhiteSpaces( item.textNode );
+				const { parent, startOffset } = item.textNode;
+
+				// Make sure the position is after all leading whitespaces in the text node.
+				const position = model.createPositionAt( parent, startOffset + leadingWhiteSpaces.length );
+
+				positions.push( position );
+			}
+		}
+	}
+
+	return positions;
+}
+
+/**
+ * Checks if any of the blocks within model selection is a code block.
+ *
+ * @param {module:engine/model/selection~Selection} selection
+ * @returns {Boolean}
+ */
+export function isModelSelectionInCodeBlock( selection ) {
+	const firstBlock = first( selection.getSelectedBlocks() );
+
+	return firstBlock && firstBlock.is( 'codeBlock' );
 }
